tests: Make given running on multi node cluster more public
Move `we are on a multi node cluster` from test registry to conftest, so
that it can be used in other tests
